Revert "Create a 'FeatureNotifications' component to use in Notifications component (#1154)"
This reverts commit 78efcdcf2f5ad788268a0861d82c7c3f6605e101.
