+++ src/components/NotificationsDropdown.tsx
@@ -1,5 +1,5 @@
 import moment from 'moment';
-import { Badge, IconButton, List, ListItem, ListItemIcon, ListItemText, Popover, Theme } from '@mui/material';
+import { Badge, Box, IconButton, List, ListItem, ListItemIcon, ListItemText, Popover, Theme } from '@mui/material';
 import React, { useCallback, useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useHistory } from 'react-router';
@@ -15,7 +15,15 @@ import stopPropagation from 'src/utils/stopPropagationEvent';
 import { makeStyles } from '@mui/styles';
 import useDeviceInfo from 'src/utils/useDeviceInfo';
 import Timestamp from './common/Timestamp';
-import useFeatureNotifications from './FeatureNotification';
+import { InitializedUnits, weightSystemsNames } from 'src/units';
+import { OrganizationService, PreferencesService, UserService } from 'src/services';
+import { useOrganization, useTimeZones, useUser } from 'src/providers';
+import isEnabled from 'src/features';
+import { InitializedTimeZone, TimeZoneDescription } from 'src/types/TimeZones';
+import { getTimeZone, getUTC } from 'src/utils/useTimeZoneUtils';
+import { getTodaysDateFormatted } from '@terraware/web-components/utils';
+import { supportedLocales } from 'src/strings/locales';
+import TextWithLink from './common/TextWithLink';
 
 interface StyleProps {
   isMobile?: boolean;
@@ -154,7 +162,87 @@ export default function NotificationsDropdown(props: NotificationsDropdownProps)
   const [lastSeen, setLastSeen] = useState<number>(0);
   const [notifications, setNotifications] = useState<Notifications>();
 
-  const featureNotifications = useFeatureNotifications();
+  const timeZoneFeatureEnabled = isEnabled('Timezones');
+  const weightUnitsEnabled = isEnabled('Weight units');
+
+  const { user, reloadUser, userPreferences, reloadUserPreferences: reloadPreferences } = useUser();
+  const { selectedOrganization, reloadOrganizations } = useOrganization();
+  const timeZones = useTimeZones();
+
+  const [unitNotification, setUnitNotification] = useState(false);
+  const [timeZoneUserNotification, setTimeZoneUserNotification] = useState(false);
+  const [timeZoneOrgNotification, setTimeZoneOrgNotification] = useState(false);
+  const [userTimeZone, setUserTimeZone] = useState<string>();
+  const [orgTimeZone, setOrgTimeZone] = useState<string>();
+
+  useEffect(() => {
+    const getDefaultTimeZone = (): TimeZoneDescription => {
+      const browserTimeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
+      return getTimeZone(timeZones, browserTimeZone) || getUTC(timeZones);
+    };
+
+    const notifyTimeZoneUpdates = (userTz: InitializedTimeZone, orgTz: InitializedTimeZone) => {
+      const notifyUser = userTz.timeZone && !userTz.timeZoneAcknowledgedOnMs;
+      const notifyOrg = orgTz.timeZone && !orgTz.timeZoneAcknowledgedOnMs;
+      setUserTimeZone(userTz.timeZone);
+      setOrgTimeZone(orgTz.timeZone);
+      setTimeZoneOrgNotification(!!notifyOrg);
+      setTimeZoneUserNotification(!!notifyUser);
+    };
+
+    const initializeTimeZones = async () => {
+      if (!user) {
+        return;
+      }
+
+      const userTz: InitializedTimeZone = await UserService.initializeTimeZone(user, getDefaultTimeZone().id);
+      if (!userTz.timeZone) {
+        return;
+      }
+
+      let orgTz: InitializedTimeZone = {};
+      orgTz = await OrganizationService.initializeTimeZone(selectedOrganization, userTz.timeZone);
+
+      if (userTz.updated) {
+        reloadUser();
+      }
+
+      if (orgTz.updated) {
+        reloadOrganizations();
+      }
+
+      if (!userTz.updated && !orgTz.updated) {
+        notifyTimeZoneUpdates(userTz, orgTz);
+      }
+    };
+
+    if (timeZoneFeatureEnabled) {
+      initializeTimeZones();
+    }
+  }, [reloadOrganizations, reloadUser, selectedOrganization, timeZoneFeatureEnabled, user, userPreferences, timeZones]);
+
+  useEffect(() => {
+    const initializeWeightUnits = async () => {
+      const userUnit: InitializedUnits = await UserService.initializeUnits('metric');
+      if (!userUnit.units) {
+        return;
+      }
+
+      if (userUnit.updated) {
+        reloadPreferences();
+      }
+
+      if (!userUnit.unitsAcknowledgedOnMs) {
+        setUnitNotification(true);
+      } else {
+        setUnitNotification(false);
+      }
+    };
+
+    if (weightUnitsEnabled) {
+      initializeWeightUnits();
+    }
+  }, [user, userPreferences, weightUnitsEnabled, reloadPreferences]);
 
   const populateNotifications = useCallback(async () => {
     const notificationsData = await getNotifications();
@@ -168,9 +256,76 @@ export default function NotificationsDropdown(props: NotificationsDropdownProps)
       return dateB.getTime() - dateA.getTime();
     });
 
-    notificationsData.items = [...featureNotifications, ...notificationsData.items];
+    if (timeZoneOrgNotification) {
+      notificationsData.items.unshift({
+        id: -2,
+        notificationCriticality: 'Info',
+        organizationId: selectedOrganization.id,
+        title: strings.REVIEW_YOUR_ORGANIZATION_SETTING,
+        body: (
+          <div>
+            <ul>
+              <li>{strings.formatString(strings.TIME_ZONE_SELECTED, orgTimeZone || '')}</li>
+            </ul>
+            <Box paddingTop={1}>
+              <TextWithLink text={strings.ORG_NOTIFICATION_ACTION} href={APP_PATHS.ORGANIZATION} />
+            </Box>
+          </div>
+        ),
+        localUrl: APP_PATHS.ORGANIZATION,
+        createdTime: getTodaysDateFormatted(),
+        isRead: false,
+        hideDate: true,
+      });
+    }
+
+    if (unitNotification || timeZoneUserNotification) {
+      notificationsData.items.unshift({
+        id: -1,
+        notificationCriticality: 'Info',
+        organizationId: selectedOrganization.id,
+        title: strings.REVIEW_YOUR_ACCOUNT_SETTING,
+        body: (
+          <Box>
+            <ul>
+              <li>
+                {strings.formatString(
+                  strings.DEFAULT_LANGUAGE_SELECTED,
+                  supportedLocales.find((sLocale) => sLocale.id === user?.locale)?.name || ''
+                )}
+              </li>
+              <li>{strings.formatString(strings.TIME_ZONE_SELECTED, userTimeZone || '')}</li>
+              <li>
+                {strings.formatString(
+                  strings.WEIGHT_SYSTEM_SELECTED,
+                  weightSystemsNames().find((ws) => ws.value === userPreferences.preferredWeightSystem)?.label || ''
+                )}
+              </li>
+            </ul>
+            <Box paddingTop={1}>
+              <TextWithLink text={strings.USER_NOTIFICATION_ACTION} href={APP_PATHS.MY_ACCOUNT} />
+            </Box>
+          </Box>
+        ),
+        localUrl: APP_PATHS.MY_ACCOUNT,
+        createdTime: getTodaysDateFormatted(),
+        isRead: false,
+        hideDate: true,
+      });
+    }
+
     setNotifications(notificationsData);
-  }, [organizationId, featureNotifications]);
+  }, [
+    organizationId,
+    timeZoneOrgNotification,
+    timeZoneUserNotification,
+    unitNotification,
+    selectedOrganization.id,
+    userTimeZone,
+    orgTimeZone,
+    user?.locale,
+    userPreferences.preferredWeightSystem,
+  ]);
 
   useEffect(() => {
     // Update notifications now.
@@ -222,11 +377,32 @@ export default function NotificationsDropdown(props: NotificationsDropdownProps)
     await populateNotifications();
   };
 
-  const markAsRead = async (read: boolean, id: number, close?: boolean, individualMarkAsRead?: () => void) => {
-    if (individualMarkAsRead) {
-      onPopoverClose();
-      individualMarkAsRead();
-      await populateNotifications();
+  const markUserNotificationAsRead = async () => {
+    onPopoverClose();
+    await PreferencesService.updateUserPreferences({
+      unitsAcknowledgedOnMs: Date.now(),
+      timeZoneAcknowledgedOnMs: Date.now(),
+    });
+
+    await reloadPreferences();
+    await populateNotifications();
+  };
+
+  const markOrgNotificationAsRead = async () => {
+    onPopoverClose();
+    await PreferencesService.updateUserOrgPreferences(selectedOrganization.id, {
+      timeZoneAcknowledgedOnMs: Date.now(),
+    });
+
+    await reloadPreferences();
+    await populateNotifications();
+  };
+
+  const markAsRead = async (read: boolean, id: number, close?: boolean) => {
+    if (id === -1) {
+      markUserNotificationAsRead();
+    } else if (id === -2) {
+      markOrgNotificationAsRead();
     } else {
       if (close) {
         onPopoverClose();
@@ -301,7 +477,7 @@ export default function NotificationsDropdown(props: NotificationsDropdownProps)
 
 type NotificationItemProps = {
   notification: Notification;
-  markAsRead: (read: boolean, id: number, close?: boolean, individualMarkAsRead?: () => void) => void;
+  markAsRead: (read: boolean, id: number, close?: boolean) => void;
   reloadOrganizationData: (selectedOrgId?: number) => void;
 };
 
@@ -318,7 +494,7 @@ function NotificationItem(props: NotificationItemProps): JSX.Element {
       const orgId: string | null = new URL(localUrl, window.location.origin).searchParams.get('organizationId');
       await reloadOrganizationData(orgId ? parseInt(orgId, 10) : undefined);
     }
-    markAsRead(read, id, close, notification.markAsRead);
+    markAsRead(read, id, close);
   };
 
   const onMouseEnter = () => {
