+++ src/types/Notifications.ts
@@ -24,7 +24,6 @@ export type Notification = {
   createdTime: string;
   isRead: boolean;
   hideDate?: boolean;
-  markAsRead?: () => void;
 };
 
 export type Notifications = {
